% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{didgpu_backend_info}
\alias{didgpu_backend_info}
\title{Report installed backends and their availability}
\usage{
didgpu_backend_info()
}
\value{
A data.frame with columns \code{backend}, \code{available}, \code{notes}.
}
\description{
Report installed backends and their availability
}
\examples{
didgpu_backend_info()
}
