% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{glance.didgpu_result}
\alias{glance.didgpu_result}
\title{broom::glance method for didgpu_result}
\usage{
glance.didgpu_result(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_result}.}

\item{...}{Unused.}
}
\value{
A one-row data.frame.
}
\description{
broom::glance method for didgpu_result
}
