% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{plot.didgpu_loo_result}
\alias{plot.didgpu_loo_result}
\title{Plot method for didgpu_loo_result (tornado plot)}
\usage{
\method{plot}{didgpu_loo_result}(x, n_show = 20L, ...)
}
\arguments{
\item{x}{A \code{didgpu_loo_result}.}

\item{n_show}{Integer. Number of top rows to show. Default \code{20L}.}

\item{...}{Extra args for plot().}
}
\value{
The input invisibly.
}
\description{
Plot method for didgpu_loo_result (tornado plot)
}
