% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{print.didgpu_bacon}
\alias{print.didgpu_bacon}
\title{Print method for didgpu_bacon}
\usage{
\method{print}{didgpu_bacon}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_bacon} object from \code{\link[=didgpu_bacon]{didgpu_bacon()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for didgpu_bacon
}
