% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{print.didgpu_cs_result}
\alias{print.didgpu_cs_result}
\title{Print method for didgpu_cs_result}
\usage{
\method{print}{didgpu_cs_result}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_cs_result}.}

\item{...}{Unused.}
}
\value{
The input invisibly.
}
\description{
Print method for didgpu_cs_result
}
