% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_continuous.R
\name{print.didgpu_did_continuous_result}
\alias{print.didgpu_did_continuous_result}
\title{Print method for didgpu_did_continuous_result}
\usage{
\method{print}{didgpu_did_continuous_result}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_did_continuous_result}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for didgpu_did_continuous_result
}
