% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence.R
\name{print.didgpu_equivalence}
\alias{print.didgpu_equivalence}
\title{Print method for didgpu_equivalence}
\usage{
\method{print}{didgpu_equivalence}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_equivalence} object from \code{\link[=didgpu_equivalence]{didgpu_equivalence()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for didgpu_equivalence
}
