% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freyaldenhoven.R
\name{print.didgpu_freyaldenhoven_result}
\alias{print.didgpu_freyaldenhoven_result}
\title{Print method for didgpu_freyaldenhoven_result}
\usage{
\method{print}{didgpu_freyaldenhoven_result}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_freyaldenhoven_result}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for didgpu_freyaldenhoven_result
}
