% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_did.R
\name{print.didgpu_honest_did_result}
\alias{print.didgpu_honest_did_result}
\title{Print method for didgpu_honest_did_result}
\usage{
\method{print}{didgpu_honest_did_result}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_honest_did_result}.}

\item{...}{Unused.}
}
\value{
The input invisibly.
}
\description{
Print method for didgpu_honest_did_result
}
