% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence.R
\name{print.didgpu_joint_placebo}
\alias{print.didgpu_joint_placebo}
\title{Print method for didgpu_joint_placebo}
\usage{
\method{print}{didgpu_joint_placebo}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_joint_placebo} object from \code{\link[=didgpu_joint_placebo]{didgpu_joint_placebo()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for didgpu_joint_placebo
}
