% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfe.R
\name{print.didgpu_twfe_result}
\alias{print.didgpu_twfe_result}
\title{Print method for didgpu_twfe_result}
\usage{
\method{print}{didgpu_twfe_result}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_twfe_result}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for didgpu_twfe_result
}
