% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.didgpu_result}
\alias{tidy.didgpu_result}
\title{broom::tidy method for didgpu_result}
\usage{
tidy.didgpu_result(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_result}.}

\item{...}{Passed to \code{\link[=didgpu_tidy]{didgpu_tidy()}}.}
}
\value{
A data.frame.
}
\description{
broom::tidy method for didgpu_result
}
