% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{print.didgpu_result}
\alias{print.didgpu_result}
\title{Print method for didgpu_result}
\usage{
\method{print}{didgpu_result}(x, ...)
}
\arguments{
\item{x}{A \code{didgpu_result} object.}

\item{...}{Unused (for S3 method compatibility).}
}
\value{
The input invisibly.
}
\description{
Print method for didgpu_result
}
